Corrected issue with compiling on 64bit.  Hackish fix.
size_t changes with architecture, on my 64bit compile it was a ulong, but the underlying libs just use int and uint.  Chose int because it can automatically be cast to uint losslessly, but the reverse isn't true.

A full correct fix would be to dig through the code replacing all non-pointer uses of size_t with the correct integer type.  That's just outside of my current time limits to fix this.