Added version DMagick_No_Display to disable the entire display subsystem.
I chose to express this in the negative so that users will, as has been the case heretofore, get the full functionality.  User who explicitly don't want the GUI operations can declare the version at compile time and it will be gone.

I also chose to leave the display functions stubbed in the no-display version as this prevents API differences: the functions will still be there, but won't operate.  Users who choose to call them will have had to expressly declare that they don't work and so _shouldn't_ be surprised that the functions don't do anything!

To be plain: adding -version=DMagick_No_Display to the compiler command will eliminate the need to link User32.lib under Windows, and cause no graphical windows to show when the example code is executed.